Removed unused variables, improve test credential handling, refresh blob metadata on OutputStream close
